Data Transfer app on Python
I would like someone to help me improve on this app I have made (it's not standalone) via CGPT. I would appreciate it if someone could run this app and test it. You would be wasting your time if you just told me the problem and how I can improve it, as I have used CGPT for everything. I wanted to create an app that would solve the problem of transferring huge games with my friends as I used Google's nearby share for PC, and it doesn't allow more than 500 files at a time (No size limit). As we all know, AAA games have more than 500 files, so it would be impossible to use it. Hence, the app solves the suffering. Honestly, I was surprised by how I could understand what mistakes I was making and how I could solve them by CGPT. I would appreciate if someone could make the changes to the code and get it to work and improve it.